refactor: extract generic permission helpers to :common:android
Move the Android-only permission utilities (hasPermission, hasAllPermissions, isExternalStorageManager(Compat), hasLegacyStorageAccessPermission, MANAGE_EXTERNAL_STORAGE) out of :AnkiDroid's Permissions object into
:common:android as top-level functions.

The remaining Permissions object in :AnkiDroid keeps Anki-specific functionality (startup permission sets, dialog interactions, settings).

- adds androidx.core.ktx as a :common:android dependency, needed for ContextCompat.checkSelfPermission.
