+++ AnkiDroid/src/main/java/com/ichi2/anki/ui/windows/permissions/PermissionsStartingAt30Fragment.kt
@@ -22,8 +22,8 @@ import android.view.ViewGroup
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.annotation.RequiresApi
 import com.ichi2.anki.R
+import com.ichi2.anki.common.permissions.MANAGE_EXTERNAL_STORAGE
 import com.ichi2.anki.databinding.FragmentPermissionsStartingAt30Binding
-import com.ichi2.utils.Permissions
 import com.ichi2.utils.Permissions.canManageExternalStorage
 
 /**
@@ -31,7 +31,7 @@ import com.ichi2.utils.Permissions.canManageExternalStorage
  * if the user [canManageExternalStorage], which isn't possible in the play store.
  *
  * Requested permissions:
- * 1. All files access: [Permissions.MANAGE_EXTERNAL_STORAGE].
+ * 1. All files access: [MANAGE_EXTERNAL_STORAGE].
  *   Used for saving the collection in a public directory
  *   which isn't deleted when the app is uninstalled
  */
